angleIncrement now parseFloat; adjusted sliders

+++ projects/euler_spiral/euler_spiral.js
@@ -74,7 +74,7 @@ startAngleSlider.addEventListener('input', (event) => {
 });
 
 angleIncrementSlider.addEventListener('input', (event) => {
-  angleIncrement = parseInt(angleIncrementSlider.value);
+  angleIncrement = parseFloat(angleIncrementSlider.value);
   drawSpiral();
 });
